Split swiftdeps handler from SwiftCompiler
